fix(ui): Global Selection values should not be "sticky" in Discover
If you do not have multiple projects, we need to keep a project selected for a better experience. If your query has a project defined, it should use that, otherwise keep the project that you have previously selected, selected.

However if you *do* have multiple projects feature, we do not want to load your last selected projects as saved queries can either have a project defined or have no project defined (in which case we want to see "My Projects").